refactor: ReadThumbnail -> Play로 이동
