+++ src/main/java/toolc/yourlist/play/infra/ReadThumbnail.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.playlist.infra;
+package toolc.yourlist.play.infra;
 
 interface ReadThumbnail {
   String find(Long playlistId);
